pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "medrax"
version = "0.1.0"
description = "medical reasoning agent for chest x-ray"
authors = [
    {name = "Adibvafa Fallahpour", email = "adibvafa.fallahpour@mail.utoronto.ca"},
]
license = {file = "LICENSE"}
requires-python = ">=3.12"
dependencies = [
    "requests>=2.25.0",
    "numpy<2",
    "langchain-core>=0.1.0",
    "langchain-community>=0.0.10",
    "langchain-openai>=0.0.2",
    "langchain-anthropic>=0.0.2",
    "langgraph>=0.0.10",
    "python-dotenv>=0.19.0",
    "pandas>=1.3.0",
    "pydantic>=1.8.0",
    "Pillow>=8.0.0",
    "torchxrayvision>=0.0.37",
    "transformers>=4.44.0",
    "tokenizers>=0.10.0",
    "sentencepiece>=0.1.95",
    "shortuuid>=1.0.0",
    "accelerate>=0.12.0",
    "peft>=0.2.0",
    "bitsandbytes>=0.35.0",
    "markdown2[all]>=2.4.0",
    "protobuf>=3.15.0",
    "scikit-learn>=0.24.0",
    "gradio>=3.0.0",
    "gradio-client>=0.2.0",
    "httpx>=0.23.0",
    "uvicorn>=0.15.0",
    "fastapi>=0.68.0",
    "einops>=0.3.0",
    "timm>=0.5.0",
    "tiktoken>=0.3.0",
    "openai>=0.27.0",
    "backoff>=1.10.0",
    "torch>=2.2.0",
    "torchvision>=0.10.0",
    "scikit-image>=0.18.0",
    "gradio>=5.0.0",
    "opencv-python>=4.8.0",
    "matplotlib>=3.8.0",
    "diffusers>=0.20.0",
    "pydicom>=2.3.0",
    "gdcm>=1.0.0",
    "pylibjpeg>=1.0.0",
    "jupyter>=1.0.0",
    "albumentations>=1.0.0",
    "pyarrow>=10.0.0",
    "tenacity>=8.0.0",
    "datasets>=2.0.0",
    "transformers[agents]>=4.44.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
    "pre-commit>=3.0.0",
    "ruff>=0.1.0",
]

test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
]

mcp = [
    "mcp>=1.0.0",
]

[project.scripts]
medrax = "main:main"
medrax-mcp = "medrax.mcp.server:main"

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "isort>=5.0.0",
    "flake8>=6.0.0",
    "pre-commit>=3.0.0",
    "ruff>=0.1.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100
select = ["E", "F", "W", "I", "N", "UP", "YTT", "ANN", "S", "BLE", "FBT", "B", "A", "COM", "C4", "DTZ", "T10", "EM", "EXE", "FA", "ISC", "ICN", "G", "INP", "PIE", "T20", "PYI", "PT", "Q", "RSE", "RET", "SLF", "SIM", "TID", "TCH", "ARG", "THR", "PTH", "ERA", "PGH", "PL", "TRY", "FLY", "NPY", "PERF", "RUF"]
ignore = ["ANN101", "ANN102", "S101", "S104", "S105", "S106"]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]
"tests/*" = ["S101", "S105", "S106"]

[tool.black]
line-length = 100
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra -q --strict-markers"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
]
